change to use StaffRepository from StaffService instead of multiple services

+++ src/main/java/com/samoonpride/backend/repository/SubscribeRepository.java
@@ -3,7 +3,7 @@ package com.samoonpride.backend.repository;
 import com.samoonpride.backend.model.Subscribe;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SubscribeRepository extends CrudRepository<Subscribe, Long> {
+public interface SubscribeRepository extends CrudRepository<Subscribe, Integer> {
     boolean existsByLineUser_UserIdAndIssueId(String lineUserId, int issueId);
     void deleteByLineUser_UserIdAndIssueId(String lineUserId, int issueId);
